+++ setup.cfg
@@ -1,5 +1,5 @@
 [zest.releaser]
-python-file-with-version = pydot.py
+python-file-with-version = src/pydot/__init__.py
 release=no
 push-changes=no
 create-wheel=yes
