+++ src/pydot/__init__.py
@@ -1,0 +1,7 @@
+"""An interface to GraphViz."""
+
+__author__ = "Ero Carrera"
+__version__ = "2.0.0.dev0"
+__license__ = "MIT"
+
+from pydot.core import *
